# server/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3010",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format **Bearer &lt;token&gt;**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/activities/{ActivityId}": {
      "get": {
        "tags": [
          "Activities"
        ],
        "description": "",
        "parameters": [
          {
            "name": "ActivityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Activity Data",
            "schema": {
              "$ref": "#/definitions/ActivityGetData"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Activities"
        ],
        "description": "",
        "parameters": [
          {
            "name": "ActivityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Content": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Activities"
        ],
        "description": "",
        "parameters": [
          {
            "name": "ActivityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/discussions/{DiscussionId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "DiscussionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "DiscussionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/events/{EventId}": {
      "get": {
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "EventId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Event Data",
            "schema": {
              "$ref": "#/definitions/EventsData"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "EventId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ActivityGetData": {
      "type": "object",
      "properties": {
        "Content": {
          "type": "string",
          "example": "This is a sample activity content."
        },
        "UserId": {
          "type": "string",
          "example": "user123"
        },
        "MediaFiles": {
          "type": "array",
          "example": [
            "https://example.com/image.jpg"
          ],
          "items": {
            "type": "string"
          }
        },
        "Documents": {
          "type": "array",
          "example": [
            "https://example.com/document.pdf"
          ],
          "items": {
            "type": "string"
          }
        },
        "Mentions": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "Username": {
                "type": "string",
                "example": "user1345"
              },
              "UserId": {
                "type": "string",
                "example": "6df8f7wed689ef"
              },
              "FullName": {
                "type": "string",
                "example": "User"
              }
            }
          }
        },
        "NoOfLikes": {
          "type": "number",
          "example": 2
        },
        "NoOfComments": {
          "type": "number",
          "example": 2
        },
        "DocId": {
          "type": "string",
          "example": "activity_1234"
        },
        "UserDetails": {
          "type": "object",
          "properties": {
            "FullName": {
              "type": "string",
              "example": "Jane Doe"
            },
            "Username": {
              "type": "string",
              "example": "janedoe"
            },
            "Email": {
              "type": "string",
              "example": "jane.doe@example.com"
            },
            "DocId": {
              "type": "string",
              "example": "abc123DEF"
            },
            "Country": {
              "type": "string",
              "example": "Canada"
            },
            "City": {
              "type": "string",
              "example": "Toronto"
            },
            "JobTitle": {
              "type": "string",
              "example": "Product Manager"
            },
            "Industry": {
              "type": "string",
              "example": "Software"
            },
            "CompanyName": {
              "type": "string",
              "example": "InnovateTech"
            },
            "About": {
              "type": "string",
              "example": "Seasoned product manager with over 10 years of experience in the software industry."
            },
            "CoverPicture": {
              "type": "string",
              "example": "string"
            },
            "ProfilePicture": {
              "type": "string",
              "example": "string"
            }
          }
        },
        "HasLiked": {
          "type": "boolean",
          "example": true
        },
        "HasSaved": {
          "type": "boolean",
          "example": false
        }
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}